webprovider: Copied code from webapi 'core'. Added support for log4j. Updated some ctor code for generics.

+++ src/com/inetvod/common/core/ISO8601DateFormat.java
@@ -4,6 +4,7 @@ package com.inetvod.common.core;
 
 import java.text.FieldPosition;
 import java.text.ParsePosition;
+import java.text.DateFormat;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.TimeZone;
@@ -15,91 +16,101 @@ import java.util.TimeZone;
  * No left truncation is allowed.
  * An optional following time zone qualifier is allowed as for dateTime.
  */
-public class ISO8601DateFormat extends ISO8601DateTimeFormat {
+public class ISO8601DateFormat extends ISO8601DateTimeFormat
+{
 
-  /**
-   * Construct a new ISO8601DateFormat using the default time zone.
-   *
-   */
-  public ISO8601DateFormat() {
-    setCalendar(Calendar.getInstance());
-  }
+	/**
+	 * Construct a new ISO8601DateFormat using the default time zone.
+	 */
+	public ISO8601DateFormat()
+	{
+		setCalendar(Calendar.getInstance());
+	}
 
-  /**
-   * Construct a new ISO8601DateFormat using a specific time zone.
-   * @param tz The time zone used to format and parse the date.
-   */
-  public ISO8601DateFormat(TimeZone tz) {
-    setCalendar(Calendar.getInstance(tz));
-  }
+	/**
+	 * Construct a new ISO8601DateFormat using a specific time zone.
+	 * @param tz The time zone used to format and parse the date.
+	 */
+	public ISO8601DateFormat(TimeZone tz)
+	{
+		setCalendar(Calendar.getInstance(tz));
+	}
 
-  /**
-   * @see java.text.DateFormat#parse(String, ParsePosition)
-   */
-  public Date parse(String text, ParsePosition pos) {
+	/**
+	 * @see DateFormat#parse(String, ParsePosition)
+	 */
+	public Date parse(String text, ParsePosition pos)
+	{
 
-    int i = pos.getIndex();
+		int i = pos.getIndex();
 
-    try {
-      int year = Integer.valueOf(text.substring(i, i + 4)).intValue();
-      i += 4;
+		try
+		{
+			int year = Integer.valueOf(text.substring(i, i + 4));
+			i += 4;
 
-      if (text.charAt(i) != '-') {
-        throw new NumberFormatException();
-      }
-      i++;
+			if(text.charAt(i) != '-')
+			{
+				throw new NumberFormatException();
+			}
+			i++;
 
-      int month = Integer.valueOf(text.substring(i, i + 2)).intValue() - 1;
-      i += 2;
+			int month = (Integer.valueOf(text.substring(i, i + 2))) - 1;
+			i += 2;
 
-      if (text.charAt(i) != '-') {
-        throw new NumberFormatException();
-      }
-      i++;
+			if(text.charAt(i) != '-')
+			{
+				throw new NumberFormatException();
+			}
+			i++;
 
-      int day = Integer.valueOf(text.substring(i, i + 2)).intValue();
-      i += 2;
+			int day = Integer.valueOf(text.substring(i, i + 2));
+			i += 2;
 
-      calendar.set(year, month, day);
-      calendar.set(Calendar.HOUR_OF_DAY, 0);
-      calendar.set(Calendar.MINUTE, 0);
-      calendar.set(Calendar.SECOND, 0);
-      calendar.set(Calendar.MILLISECOND, 0); // no parts of a second
+			calendar.set(year, month, day);
+			calendar.set(Calendar.HOUR_OF_DAY, 0);
+			calendar.set(Calendar.MINUTE, 0);
+			calendar.set(Calendar.SECOND, 0);
+			calendar.set(Calendar.MILLISECOND, 0); // no parts of a second
 
-      i = parseTZ(i, text);
+			i = parseTZ(i, text);
 
-    }
-    catch (NumberFormatException ex) {
-      pos.setErrorIndex(i);
-      i = 0;
-      return null;
-    }
-    catch (IndexOutOfBoundsException ex) {
-      pos.setErrorIndex(i);
-      i = 0;
-      return null;
-    }
-    finally {
-      pos.setIndex(i);
-    }
+		}
+		catch(NumberFormatException ex)
+		{
+			pos.setErrorIndex(i);
+			i = 0;
+			return null;
+		}
+		catch(IndexOutOfBoundsException ex)
+		{
+			pos.setErrorIndex(i);
+			i = 0;
+			return null;
+		}
+		finally
+		{
+			pos.setIndex(i);
+		}
 
-    return calendar.getTime();
-  }
+		return calendar.getTime();
+	}
 
-  /**
-   * @see java.text.DateFormat#format(Date, StringBuffer, FieldPosition)
-   */
-  public StringBuffer format(
-    Date date,
-    StringBuffer sbuf,
-    FieldPosition fieldPosition) {
+	/**
+	 * @see DateFormat#format(Date, StringBuffer, FieldPosition)
+	 */
+	public StringBuffer format(
+		Date date,
+		StringBuffer sbuf,
+		FieldPosition fieldPosition)
+	{
 
-    calendar.setTime(date);
+		calendar.setTime(date);
 
-    writeCCYYMM(sbuf);
+		writeCCYYMM(sbuf);
 
-    //writeTZ(sbuf);
+		//writeTZ(sbuf);
 
-    return sbuf;
-  }
+		return sbuf;
+	}
 }
